chore: cleanup for super-this merge
