+++ src/blockstore/index.ts
@@ -14,4 +14,4 @@ export { EncryptedBlockstore, BaseBlockstore, CompactionFetcher, type BlockFetch
 export { Loader } from "./loader.js";
 export { parseCarFile } from "./loader-helpers.js";
 export { ConnectionBase } from "./connection-base.js";
-export { setCryptoKeyFromGatewayMetaPayload, addCryptoKeyToGatewayMetaPayload } from "./store-meta.js";
+export { setCryptoKeyFromGatewayMetaPayload, addCryptoKeyToGatewayMetaPayload } from "./meta-key-helper.js";
